% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_names.R
\name{cat_names}
\alias{cat_names}
\title{A function to print variable names for easy pasting to a new character vector.}
\usage{
cat_names(x)
}
\arguments{
\item{x}{a data.frame}
}
\description{
A function to print variable names for easy pasting to a new character vector.
}
\examples{
cat_names(iris)

}
