% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportGBMtoPDF.R
\name{exportGBMtoPDF}
\alias{exportGBMtoPDF}
\title{Function to write the variable importance plots of a gbm object to a pdf.}
\usage{
exportGBMtoPDF(gbm_object, file_path, file_name)
}
\arguments{
\item{gbm_object}{An object containing a GBM model and components}

\item{file_path}{A character string pointing to the directory of where to write the PDF containing the partial dependency plots.}

\item{file_name}{A character string pointing to the file of where to write the PDF containing the partial dependency plots.}
}
\value{
NA - writes to file.
}
\description{
Function to write the variable importance plots of a gbm object to a pdf.
}
\examples{
exportGBMtoPDF(gbm1, '/data/output/', 'output.pdf')


}
\author{
Arshan Tarapore
\email{aktarapore@gmail.com}
}
