% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiFreqFact.R
\name{hiFreqFact}
\alias{hiFreqFact}
\title{Reduce Cardinality of Factor Variables to only keep Hi Frequency Levels}
\usage{
hiFreqFact(x, y = 1000)
}
\arguments{
\item{x}{- vector of factor values}

\item{y}{- minimum number of occurrences to keep}
}
\description{
Reduce Cardinality of Factor Variables to only keep Hi Frequency Levels
}
