% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.objects.R
\name{lsObjects}
\alias{lsObjects}
\title{Hidden function to list objects and how big they are}
\usage{
lsObjects(pos = 1, pattern, order.by, decreasing = FALSE, head = FALSE,
  n = 5)
}
\arguments{
\item{pos}{TODO: DOCUMENT ME!}

\item{pattern}{TODO: DOCUMENT ME!}

\item{order.by}{TODO: DOCUMENT ME!}

\item{decreasing}{TODO: DOCUMENT ME!}

\item{head}{TODO: DOCUMENT ME!}

\item{n}{TODO: DOCUMENT ME!}
}
\description{
This object lists objects by size
}
