% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lsos}
\alias{lsos}
\title{A Memory Object Browser Function
Shamelessly Copied from this stackoverflow.com post
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
\usage{
lsos(..., n = 10)
}
\description{
A Memory Object Browser Function
Shamelessly Copied from this stackoverflow.com post
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
}
\examples{
lsos()
}
\keyword{memory}

