% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/naHandler.R
\name{naHandler}
\alias{naHandler}
\title{A Function to handle NA values.  If column is a factor variable, the NA's are
converted to a separate factor level.  If column is numeric, the mean is imputed.}
\usage{
naHandler(x)
}
\arguments{
\item{x}{A vector}
}
\description{
A Function to handle NA values.  If column is a factor variable, the NA's are
converted to a separate factor level.  If column is numeric, the mean is imputed.
}
\examples{
iris[,c('Petal.Length','Petal.Width','Species')] <- lapply(iris[,c('Petal.Length','Petal.Width','Species')], function(x) naHandler(x))
}
\keyword{NA}
\keyword{impute}
\keyword{mean}

