% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileDat.R
\name{prof_dat}
\alias{prof_dat}
\title{Standard method of profiling a data frame}
\usage{
prof_dat(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
data.frame with profiling data from input data.frame
}
\description{
Standard method of profiling a data frame
}
\examples{
prof_dat(iris)

}
