% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTypes.R
\name{setTypes}
\alias{setTypes}
\title{Set the variable types for a data.frame}
\usage{
setTypes(x)
}
\arguments{
\item{x}{}
}
\description{
Set the variable types for a data.frame
}
\examples{
setTypes(iris)
}
