% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cnn}
\alias{cnn}
\title{A Function to print the column number, name and variable type for each column in a data.frame.}
\usage{
cnn(dataset)
}
\arguments{
\item{x}{A data.frame object}
}
\description{
This function prints the column number, name, and variable type for each column in a data.frame.
}
\examples{
cnn(iris)
}
\keyword{class}
\keyword{names}

