% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date.R
\name{convert_date}
\alias{convert_date}
\title{Converts all columns of class POSIXct or POSIXt in a data.frame to Date class.}
\usage{
convert_date(xx)
}
\arguments{
\item{xx}{A data.frame containing columns of class POSIXct or POSIXt}
}
\value{
data.frame with any columns of class POSIXct or POSIXt converted to Dates
}
\description{
Converts all columns of class POSIXct or POSIXt in a data.frame to Date class.
}
\examples{
df2 <- convert_date(df1)


}
\author{
\itemize{
   \item Andrei Rukavina - \url{https://github.com/arukavina}
   \item Thijn van der Heijden - \email{avanderheijden@alixpartners.com}
   \item Zach Armentrout - \email{zarmentrout@alixpartners.com}
   \item Qianbo Wang - \email{qwang@alixpartners.com}
   \item James Wang - \email{swang@alixpartners.com}
}
}
