% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsos.R
\name{lsos}
\alias{lsos}
\title{A Memory Object Browser Function
Shamelessly Copied from this stackoverflow.com post
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{...}{Passed to .ls.objects}

\item{n}{Number of objects to show}
}
\description{
A Memory Object Browser Function
Shamelessly Copied from this stackoverflow.com post
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
}
\examples{
data(iris)
lsos()
}
\keyword{memory}

