% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanImp.R
\name{meanImp}
\alias{meanImp}
\title{A Function to impute means from all other non-NA cells in a column.}
\usage{
meanImp(x)
}
\arguments{
\item{x}{A vector}
}
\description{
A Function to impute means from all other non-NA cells in a column.
}
\examples{
meanImp(iris$Petal.Length)
}
\keyword{impute}
\keyword{mean}

